#15 extract grid styles
Grid styles are now applied separately, so they do not receive the global focus styling from general.css.

The "http(s)?://.*/tas/secure/.*?action=.*" regex is still part of the big bulk of styling, as this is the only way to match a natural selection page, a more specific regex seems to fail.